query fix

+++ tpcds-spark.ipynb.py
@@ -233,7 +233,7 @@ def run_query(run_id, query_number, queries, path_to_save_results, data_size, pr
             elapsed_time = end - start
             execution_times.append(elapsed_time)
         
-        elapsed_time = np.mean(execution_times)
+        elapsed_time = float(np.mean(execution_times))
         result.write.format("csv").mode("overwrite").option("header", "true").save(path_to_save_results.format(size=data_size, query_number=query_number))
         
         stats = {
